+++ src/Distributions/Uniform.cpp
@@ -3,7 +3,7 @@
 
 #include "Distribution.h"
 
-typedef Vec3<double> vec3;
+typedef Vec3<float> vec3;
 
 UniformDistribution::UniformDistribution() : Distribution() {}
 
@@ -17,6 +17,6 @@ vec3 UniformDistribution::sample(const vec3 &pos, const vec3 &n) {
     return res;
 }
 
-double UniformDistribution::pdf(const vec3 &pos, const vec3 &n, const vec3 &d) const {
+float UniformDistribution::pdf(const vec3 &pos, const vec3 &n, const vec3 &d) const {
     return (d % n > 0 ? 1 / (M_PI * 2) : 0);
 }
