+++ src/Object/Geometry/Plane.cpp
@@ -4,15 +4,15 @@
 
 #include <vector>
 
-Plane::Plane(const Vec3<double> &v) : norm(v.norm()) {}
+Plane::Plane(const Vec3<float> &v) : norm(v.norm()) {}
 
-Vec3<double> Plane::normal(const Vec3<double> &p) const {
+Vec3<float> Plane::normal(const Vec3<float> &p) const {
     return norm;
 }
 
-std::vector<double> Plane::get_intersect_(const Ray& ray) const {
-    double t = -(ray.start % norm) / (ray.v % norm);
-    std::vector<double> res;
+std::vector<float> Plane::get_intersect_(const Ray& ray) const {
+    float t = -(ray.start % norm) / (ray.v % norm);
+    std::vector<float> res;
     if (t > 0) res.push_back(t);
     return res;
 };
